move file test
