+++ conans/test/functional/util/tools_test.py
@@ -1,6 +1,8 @@
 # -*- coding: utf-8 -*-
+
 import os
 import platform
+import subprocess
 import unittest
 import warnings
 
@@ -14,7 +16,61 @@ from conans.client.tools.win import vswhere
 from conans.errors import ConanException
 from conans.model.settings import Settings
 from conans.test.utils.mocks import TestBufferConanOutput
+from conans.test.utils.test_files import temp_folder
 from conans.util.env_reader import get_env
+from conans.util.files import save
+from conans.util.runners import check_output_runner
+
+
+class FunctionalToolsTest(unittest.TestCase):
+    output = TestBufferConanOutput()
+
+    @pytest.mark.tool_file  # Needs the "file" command, not by default in linux
+    @unittest.skipIf(which("file") is None, "Needs the 'file' command, not by default in linux")
+    def test_unix_to_dos_unit(self):
+        def save_file(contents):
+            tmp = temp_folder()
+            filepath = os.path.join(tmp, "a_file.txt")
+            save(filepath, contents)
+            return filepath
+
+        fp = save_file(b"a line\notherline\n")
+        if platform.system() != "Windows":
+            output = check_output_runner(["file", fp], stderr=subprocess.STDOUT)
+            self.assertIn("ASCII text", str(output))
+            self.assertNotIn("CRLF", str(output))
+
+            tools.unix2dos(fp)
+            output = check_output_runner(["file", fp], stderr=subprocess.STDOUT)
+            self.assertIn("ASCII text", str(output))
+            self.assertIn("CRLF", str(output))
+        else:
+            fc = tools.load(fp)
+            self.assertNotIn("\r\n", fc)
+            tools.unix2dos(fp)
+            fc = tools.load(fp)
+            self.assertIn("\r\n", fc)
+
+        self.assertEqual("a line\r\notherline\r\n", str(tools.load(fp)))
+
+        fp = save_file(b"a line\r\notherline\r\n")
+        if platform.system() != "Windows":
+            output = check_output_runner(["file", fp], stderr=subprocess.STDOUT)
+            self.assertIn("ASCII text", str(output))
+            self.assertIn("CRLF", str(output))
+
+            tools.dos2unix(fp)
+            output = check_output_runner(["file", fp], stderr=subprocess.STDOUT)
+            self.assertIn("ASCII text", str(output))
+            self.assertNotIn("CRLF", str(output))
+        else:
+            fc = tools.load(fp)
+            self.assertIn("\r\n", fc)
+            tools.dos2unix(fp)
+            fc = tools.load(fp)
+            self.assertNotIn("\r\n", fc)
+
+        self.assertEqual("a line\notherline\n", str(tools.load(fp)))
 
 
 @unittest.skipUnless(platform.system() == "Windows", "Requires Visual Studio")
